add desert and sea

+++ Assets/FishManShaderTutorial/Projects/ShaderLibs/MergeRayMarch.cginc
@@ -47,9 +47,8 @@ v2f VertMergeRayMarch(appdata_img v) {
 	return o;
 }
 
-
-fixed4 ProcessRayMarch(float2 uv,float3 ro,float3 rd,inout float sceneDep,float4 sceneCol);
-fixed4 FragMergeRayMarch(v2f i) : SV_Target{
+float4 ProcessRayMarch(float2 uv,float3 ro,float3 rd,inout float sceneDep,float4 sceneCol);
+float4 FragMergeRayMarch(v2f i) : SV_Target{
 	float depth = LinearEyeDepth(SAMPLE_DEPTH_TEXTURE(_CameraDepthTexture, i.uv_depth));
 	depth *= length(i.interpolatedRay.xyz);
 	fixed4 sceneCol = tex2D(_MainTex, i.uv);
